+++ project/activity-processor/src/main/resources/application.yml
@@ -1,7 +1,7 @@
 spring:
 
   rabbitmq:
-    addresses: rabbitmq
+    addresses: ${RABBITMQ_ADDRESS:localhost}
     
   datasource:    
     driverClassName: ${DB_DRIVER:org.h2.Driver}
